A GitHub action to apply Prettier fixes (#1)
### How to Use It

To automate Prettier, simply add [the provided YAML configuration](.github/workflows/prettier-fix.yml) to your repository. If needed, a new branch with a name starting with `prettier/` will be created, making it easy to review and merge the fixes into your main branch.

### Important Note for Workflow YAMLs

If the changes involve files within the `.github/workflows/` directory, these need to be updated manually. After making the necessary adjustments, simply create a pull request to finalize the changes and keep your repository in top shape.